Fix italics in Dina font
One of the most awaited commits in this repository.

Using "M-x describe-font" (not face) and searching for "Dina", I was able to see
that there was a line:

    -unknown-Dina-regular-normal-normal-*-8-*-*-*-c-*-iso10646-1

And another:

    -unknown-Dina-medium-italic-normal-*-10-*-*-*-c-*-iso10646-1

But that there was no "Dina-regular-italic" line. Therefore, I tried setting the
weight to "medium" instead of "regular", and Emacs was able to find the font. I
assume that other programs (like Suckless' ST) use a different method/library
for finding the italic font.