Add x8dcc/org-custom-id-make-unique, use in x8dcc/org-custom-id-get
Also add `x8dcc/org-custom-id-exists-p'.

Now if there are three headings called "foo", the ID of the first one will be
"foo", and the ID of the second one will be "foo1", and the ID of the third one
will be "foo2".

It's tempting to append numbers relevant to the context (e.g. section number),
but that would make it more unstable after buffer modifications.