Hack 'org-latex--inline-image' to wrap SVG exports in '\adjustbox{...}'
Also clear 'org-latex-image-default-width'.

This commit declares a function for modifying exported LaTeX fragments
that include SVG images, so that they are wrapped in:

  \adjustbox{max width=\linewidth}{...}

This way, since 'org-latex-image-default-width' is cleared, there is no
minimum width for SVG images (which would upscale small images occupy
the whole width of the page), while keeping the page width as the
maximum (so they don't appear on the margins).

This function is hooked to 'org-latex--inline-image' (hence the "hack"
term, specially since it's an internal function), so that it modifies
the original Org LaTeX export. Specifically, it replaces the relevant
part to turn:

  \begin{center}
  \includesvg{../img/foobar1}
  \label{fig1}
  \end{center}

Into:

  \begin{center}
  \adjustbox{max width=\linewidth}{\includesvg{../img/foobar1}}
  \label{fig1}
  \end{center}

This makes my blog PDF exports extremely beautiful, if I may say so
myself.